requirements.txt
# CoMIDF Requirements
# Python dependencies for Edge Agent and Cloud Platform

# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
numpy>=1.24.0

# FastAPI for Cloud Platform API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# gRPC for secure communication
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Security
PyJWT>=2.8.0
cryptography>=41.0.0

# Database (optional)
psycopg2-binary>=2.9.9
redis>=5.0.0

# Kafka for event streaming (optional)
kafka-python>=2.0.2

# Monitoring and logging
prometheus-client>=0.19.0

# Network utilities
scapy>=2.5.0
netifaces>=0.11.0

# LLM integration (optional)
openai>=1.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
mypy>=1.7.0

